+++ lib/src/utils.dart
@@ -139,6 +139,16 @@ StreamSink mapSink(StreamSink original, fn(event)) {
   return controller.sink;
 }
 
+/// Like [runZoned], but [zoneValues] are set for the callbacks in
+/// [zoneSpecification] and [onError].
+runZonedWithValues(Function body(), {Map zoneValues,
+    ZoneSpecification zoneSpecification, Function onError}) {
+  return runZoned(() {
+    return runZoned(body,
+        zoneSpecification: zoneSpecification, onError: onError);
+  }, zoneValues: zoneValues);
+}
+
 /// Truncates [text] to fit within [maxLength].
 ///
 /// This will try to truncate along word boundaries and preserve words both at